\name{call.option.discount}
\alias{call.option.discount}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Discounted Option Price }
\description{
  Calculates the discounted call option price
}
\usage{
call.option.discount(payoff, T, t, r)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{payoff}{ the payoff of the call option }
  \item{T}{ time to maturity }
  \item{t}{ the current time }
  \item{r}{ the risk-free interest rate }
}
\author{ Dominik Dahlem }
\examples{
##---- Should be DIRECTLY executable !! ----
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ Call Option }% __ONLY ONE__ keyword per line
