\name{call.payoff}
\alias{call.payoff}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Payoff of a call option }
\description{
  Calculate the payoff of an asian call option
}
\usage{
call.payoff(Savg, K)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Savg}{ the average share price to be taken into account }
  \item{K}{ the strike price of the option }
}
\author{ Dominik Dahlem }
\examples{
##---- Should be DIRECTLY executable !! ----
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ call payoff }
\keyword{ asian option }% __ONLY ONE__ keyword per line
