\name{european.d1}
\alias{european.d1}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ d1 function for european option pricing }
\description{
  d1 function for european option pricing
}
\usage{
european.d1(S0, K, r, sigma, T, t0)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{S0}{ the share price at $t=0$ }
  \item{K}{ the strike price }
  \item{r}{ the risk-less interest rate }
  \item{sigma}{ the volatility }
  \item{T}{ the maturity time }
  \item{t0}{ the current time }
}
\author{ Dominik Dahlem }
\examples{
##---- Should be DIRECTLY executable !! ----
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ European option }% __ONLY ONE__ keyword per line
