\name{price.dev}
\alias{price.dev}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Price Evolution }
\description{
  Evolve the share price one step.
}
\usage{
price.dev(Sold, r, sigma, rnd)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Sold}{ the current share price }
  \item{r}{ the risk-free interest rate }
  \item{sigma}{ the underlying volatility }
  \item{rnd}{ a normal random number }
}
\author{ Dominik Dahlem }
\examples{
##---- Should be DIRECTLY executable !! ----
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ share price }
\keyword{ lognormal  }% __ONLY ONE__ keyword per line
